requirements.txt
Deprecated==1.2.13
googleapis-common-protos==1.56.4
grpcio==1.47.0
opentelemetry-api==1.11.1
opentelemetry-exporter-jaeger==1.11.1
opentelemetry-exporter-jaeger-proto-grpc==1.11.1
opentelemetry-exporter-jaeger-thrift==1.11.1
opentelemetry-opentracing-shim==0.32b0
opentelemetry-sdk==1.11.1
opentelemetry-semantic-conventions==0.30b1
opentracing==2.4.0
protobuf==4.21.4
six==1.16.0
thrift==0.16.0
typing-extensions==4.3.0
wrapt==1.14.1
